+++ clients/egui/src/account/tabs/svg_editor/main.rs
@@ -1,5 +1,4 @@
 use std::borrow::BorrowMut;
-use std::f32::consts::PI;
 use std::fmt::Write;
 use std::sync::mpsc;
 
@@ -8,12 +7,10 @@ use minidom::Element;
 use resvg::tiny_skia::{Path, PathBuilder, PathSegment, Pixmap};
 use resvg::usvg::{self, Size, TreeWriting};
 
-use crate::theme::{Icon, ThemePalette};
-use crate::widgets::Button;
-const ICON_SIZE: f32 = 30.0;
-const COLOR_SWATCH_BTN_RADIUS: f32 = 9.0;
-const THICKNESS_BTN_X_MARGIN: f32 = 5.0;
-const THICKNESS_BTN_WIDTH: f32 = 30.0;
+use crate::theme::ThemePalette;
+
+use super::toolbar::{ColorSwatch, Toolbar, Component};
+
 const INITIAL_SVG_CONTENT: &str = "<svg xmlns=\"http://www.w3.org/2000/svg\" ></svg>";
 const ZOOM_STOP: f32 = 0.1;
 
@@ -25,15 +22,9 @@ pub struct SVGEditor {
     path_builder: PathBuilder,
     zoom_factor: f32,
     id_counter: usize,
-    toolbar: Toolbar,
+    pub toolbar: Toolbar,
     inner_rect: egui::Rect,
     sao_offset: egui::Vec2,
-}
-
-struct Toolbar {
-    components: Vec<Component>,
-    active_color: Option<ColorSwatch>,
-    active_stroke_width: u32,
 }
 
 impl SVGEditor {
@@ -47,42 +38,6 @@ impl SVGEditor {
         }
         let root: Element = content.parse().unwrap();
 
-        let components = vec![
-            Component::Button(SimpleButton {
-                icon: Icon::UNDO,
-                margin: egui::Margin::symmetric(4.0, 7.0),
-                coming_soon_text: Some(
-                    "Undo/Redo will be added in the next version. Stay Tuned!".to_string(),
-                ),
-            }),
-            Component::Button(SimpleButton {
-                icon: Icon::REDO,
-                margin: egui::Margin::symmetric(4.0, 7.0),
-                coming_soon_text: Some(
-                    "Undo/Redo will be added in the next version. Stay Tuned!".to_string(),
-                ),
-            }),
-            Component::Separator(egui::Margin::symmetric(10.0, 0.0)),
-            Component::Button(SimpleButton {
-                icon: Icon::BRUSH,
-                coming_soon_text: None,
-                margin: egui::Margin::symmetric(4.0, 7.0),
-            }),
-            Component::Button(SimpleButton {
-                icon: Icon::ERASER,
-                coming_soon_text: Some(
-                    "Eraser will be added in the next version. Stay Tuned!".to_string(),
-                ),
-                margin: egui::Margin::symmetric(4.0, 7.0),
-            }),
-            Component::Separator(egui::Margin::symmetric(10.0, 0.0)),
-            Component::StrokeWidth(3),
-            Component::StrokeWidth(6),
-            Component::StrokeWidth(9),
-            Component::Separator(egui::Margin::symmetric(10.0, 0.0)),
-        ];
-
-        let toolbar = Toolbar { components, active_color: None, active_stroke_width: 3 };
         let max_id = root
             .children()
             .map(|el| {
@@ -99,7 +54,7 @@ impl SVGEditor {
             draw_tx,
             id_counter: max_id as usize,
             root,
-            toolbar,
+            toolbar: Toolbar::new(),
             sao_offset: egui::vec2(0.0, 0.0),
             path_builder: PathBuilder::new(),
             inner_rect: egui::Rect::NOTHING,
@@ -337,164 +292,4 @@ fn get_path_data(path: Path) -> String {
 
     s.pop(); // ' '
     s
-}
-
-#[derive(Clone)]
-enum Component {
-    Button(SimpleButton),
-    ColorSwatch(ColorSwatch),
-    StrokeWidth(u32),
-    Separator(egui::Margin),
-}
-#[derive(Clone)]
-struct SimpleButton {
-    icon: Icon,
-    margin: egui::Margin,
-    coming_soon_text: Option<String>,
-}
-#[derive(Clone)]
-struct ColorSwatch {
-    id: String,
-    color: egui::Color32,
-}
-
-trait SizableComponent {
-    fn get_width(&self) -> f32;
-}
-impl SizableComponent for Component {
-    fn get_width(&self) -> f32 {
-        match self {
-            Component::Button(btn) => btn.margin.sum().x + ICON_SIZE,
-            Component::Separator(margin) => margin.sum().x,
-            Component::ColorSwatch(_color_btn) => COLOR_SWATCH_BTN_RADIUS * PI,
-            Component::StrokeWidth(_) => THICKNESS_BTN_WIDTH + THICKNESS_BTN_X_MARGIN * 2.0,
-        }
-    }
-}
-
-impl Toolbar {
-    fn width(&self) -> f32 {
-        self.components.iter().map(|c| c.get_width()).sum()
-    }
-    fn calculate_rect(&self, ui: &mut egui::Ui) -> egui::Rect {
-        let height = 0.0;
-        let available_rect = ui.available_rect_before_wrap();
-
-        let maximized_min_x = (available_rect.width() - self.width()) / 2.0 + available_rect.left();
-
-        let min_pos = egui::Pos2 { x: maximized_min_x, y: available_rect.top() + height };
-
-        let maximized_max_x =
-            available_rect.right() - (available_rect.width() - self.width()) / 2.0;
-
-        let max_pos = egui::Pos2 { x: maximized_max_x, y: available_rect.top() };
-        egui::Rect { min: min_pos, max: max_pos }
-    }
-
-    fn show(&mut self, ui: &mut egui::Ui) {
-        let rect = self.calculate_rect(ui);
-
-        ui.allocate_ui_at_rect(rect, |ui| {
-            ui.horizontal(|ui| {
-                self.components.iter().for_each(|c| match c {
-                    Component::Button(btn) => {
-                        egui::Frame::default()
-                            .inner_margin(btn.margin)
-                            .show(ui, |ui| {
-                                let btn_res = Button::default().icon(&btn.icon).show(ui);
-
-                                if let Some(tooltip_text) = &btn.coming_soon_text {
-                                    btn_res.on_hover_text(tooltip_text);
-                                }
-                            });
-                    }
-                    Component::Separator(margin) => {
-                        ui.add_space(margin.right);
-                        ui.add(egui::Separator::default().shrink(ui.available_height() * 0.3));
-                        ui.add_space(margin.left);
-                    }
-                    Component::ColorSwatch(btn) => {
-                        let (response, painter) = ui.allocate_painter(
-                            egui::vec2(COLOR_SWATCH_BTN_RADIUS * PI, ui.available_height()),
-                            egui::Sense::click(),
-                        );
-                        if response.clicked() {
-                            self.active_color =
-                                Some(ColorSwatch { id: btn.id.clone(), color: btn.color });
-                        }
-                        if let Some(active_color) = &self.active_color {
-                            let opacity = if active_color.id.eq(&btn.id) {
-                                1.0
-                            } else if response.hovered() {
-                                ui.output_mut(|w| w.cursor_icon = egui::CursorIcon::PointingHand);
-                                0.9
-                            } else {
-                                0.5
-                            };
-
-                            if active_color.id.eq(&btn.id) {
-                                painter.rect_filled(
-                                    response.rect.shrink2(egui::vec2(0.0, 5.0)),
-                                    egui::Rounding::same(8.0),
-                                    btn.color.gamma_multiply(0.2),
-                                )
-                            }
-                            painter.circle_filled(
-                                response.rect.center(),
-                                COLOR_SWATCH_BTN_RADIUS,
-                                btn.color.gamma_multiply(opacity),
-                            );
-                        }
-                    }
-                    Component::StrokeWidth(thickness) => {
-                        ui.add_space(THICKNESS_BTN_X_MARGIN);
-                        let (response, painter) = ui.allocate_painter(
-                            egui::vec2(THICKNESS_BTN_WIDTH, ui.available_height()),
-                            egui::Sense::click(),
-                        );
-
-                        let rect = egui::Rect {
-                            min: egui::Pos2 {
-                                x: response.rect.left(),
-                                y: response.rect.center().y - (*thickness as f32 / 3.0),
-                            },
-                            max: egui::Pos2 {
-                                x: response.rect.right(),
-                                y: response.rect.center().y + (*thickness as f32 / 3.0),
-                            },
-                        };
-
-                        if thickness.eq(&self.active_stroke_width) {
-                            painter.rect_filled(
-                                response.rect.shrink2(egui::vec2(0.0, 5.0)),
-                                egui::Rounding::same(8.0),
-                                egui::Color32::GRAY.gamma_multiply(0.1),
-                            )
-                        }
-                        if response.clicked() {
-                            self.active_stroke_width = *thickness;
-                        }
-                        if response.hovered() {
-                            ui.output_mut(|w| w.cursor_icon = egui::CursorIcon::PointingHand);
-                        }
-
-                        painter.rect_filled(
-                            rect,
-                            egui::Rounding::same(2.0),
-                            ui.visuals().text_color().gamma_multiply(0.8),
-                        );
-                        ui.add_space(THICKNESS_BTN_X_MARGIN);
-                    }
-                });
-            });
-        });
-        ui.visuals_mut().widgets.noninteractive.bg_stroke.color = ui
-            .visuals()
-            .widgets
-            .noninteractive
-            .bg_stroke
-            .color
-            .gamma_multiply(0.4);
-        ui.separator();
-    }
 }
