+++ clients/egui/src/account/tabs/svg_editor/mod.rs
@@ -1,0 +1,5 @@
+mod main;
+mod toolbar;
+
+
+pub use main::SVGEditor;
